+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Fixed: Removed shadow when elevated is zero.
+- Fixed: Border filling all area when background is transparent.
+
 ## 0.0.2
 
 - updated example.